Move more files to TypeScript
